fix(email): updated email service

+++ services/email/internal/core/domain/email.go
@@ -28,15 +28,15 @@ type SubscriptionCreatedEvent struct {
 }
 
 type WeatherUpdateEmailRequest struct {
-	To               string `json:"to" validate:"required,email"`
-	Subject          string `json:"subject" validate:"required"`
-	Name             string `json:"name" validate:"required"`
-	City             string `json:"city" validate:"required"`
-	Temperature      int    `json:"temperature"`
-	Description      string `json:"description" validate:"required"`
-	Humidity         int    `json:"humidity"`
-	WindSpeed        int    `json:"windSpeed"`
-	UnsubscribeToken string `json:"unsubscribeToken"`
+	To               string  `json:"to" validate:"required,email"`
+	Subject          string  `json:"subject" validate:"required"`
+	Name             string  `json:"name" validate:"required"`
+	City             string  `json:"city" validate:"required"`
+	Temperature      float64 `json:"temperature"`
+	Description      string  `json:"description" validate:"required"`
+	Humidity         float64 `json:"humidity"`
+	WindSpeed        float64 `json:"windSpeed"`
+	UnsubscribeToken string  `json:"unsubscribeToken"`
 }
 
 type EmailDeliveryStatus string
